fix types

+++ src/dsp/sig_source.c
@@ -39,10 +39,10 @@ int sig_source_create(float amplitude, uint64_t rx_sampling_freq, uint32_t max_o
 }
 
 void sig_source_process(int64_t freq, size_t expected_output_len, float complex **output, size_t *output_len, sig_source *source) {
-    float adjusted_freq = freq / (float) source->rx_sampling_freq;
+    double adjusted_freq = (double) freq / (double) source->rx_sampling_freq;
     for (size_t i = 0; i < expected_output_len; i++) {
         source->output[i] = cosf(source->phase) * source->amplitude + sinf(source->phase) * source->amplitude * I;
-        source->phase += 2 * M_PI * adjusted_freq;
+        source->phase += (float) (2 * M_PI * adjusted_freq);
     }
 
     *output = source->output;
